refactor: move information button
