+++ libs/journeys/ui/src/components/StepHeader/InformationButton/index.ts
@@ -1,0 +1,1 @@
+export { InformationButton } from './InformationButton'
